app.py

+++ app.py
@@ -61,7 +61,7 @@ def recommend(movie):
         st.error(f"Movie '{movie}' not found in database.")
         return [], []
     movie_index = movies[movies['title']==movie].index[0]
-    distances = similarity[movie_index]
+    distances = similarity[movie_index].astype(float)
     movies_list = sorted(list(enumerate(distances)), reverse=True, key=lambda x: x[1])[1:6]
 
     recommended_movies = []
